+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to the full browser extension will be documented in this fil
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Added
+- Report elements that have cursor pointer style.
+
 ## 0.1.8 - 2025-12-12
 
 ### Changed